[Major] seperating download.py into two files core and update, update handling the high level update coroutine and core being the subparts of the update itself. Each core function now have an async counterpart and update is now an async function
